리그 통계 버그 수정

+++ src/main/java/kr/webgori/lolien/discord/bot/dto/customgame/statistics/MostVisionScoreDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class MostVisionScoreDto {
   private long gameId;
   private String summonerName;
-  private long visionScore;
+  private float visionScore;
 }
